flasktasks/requirements.txt
apispec==3.3.1
apispec-webframeworks==0.5.2
flasgger==0.9.5
Flask==1.1.2
Flask-Celery==2.4.3
Flask-MongoEngine==0.9.5
Flask-RESTful==0.3.8
gunicorn==20.0.4
marshmallow==3.6.0
redis==3.5.3
